requirements.txt
fastapi==0.110.0
passlib==1.7.4
pydantic==2.6.4
pytest==8.1.1
python-dotenv==1.0.1
python_jose==3.3.0
SQLAlchemy==2.0.28
